+++ tests/projects/package.json/empty.js
@@ -1,0 +1,1 @@
+// retain `package.json` directory for unit test